got it working finally

+++ sirO/src/sir.rs
@@ -1,26 +1,26 @@
 use pyo3::prelude::*;
 
 pub struct Parameters {
-    pub Beta: Vec<f32>,
-    pub Gamma: Vec<f32>,
-    pub Alpha: Vec<f32>,
+    pub Beta: Vec<f64>,
+    pub Gamma: Vec<f64>,
+    pub Alpha: Vec<f64>,
 }
 
 pub struct SIR {
-    pub S: f32,
-    pub I: f32,
-    pub R: f32,
+    pub S: f64,
+    pub I: f64,
+    pub R: f64,
 }
 
 #[pyclass]
 pub struct Results {
-    pub S: Vec<f32>,
-    pub I: Vec<f32>,
-    pub R: Vec<f32>,
+    pub S: Vec<f64>,
+    pub I: Vec<f64>,
+    pub R: Vec<f64>,
 }
 
 impl Parameters {
-    pub fn new(Alpha: Vec<f32>, Beta:Vec<f32>, Gamma:Vec<f32>) -> Self {
+    pub fn new(Alpha: Vec<f64>, Beta:Vec<f64>, Gamma:Vec<f64>) -> Self {
         return Parameters {
             Alpha, Beta, Gamma
         }
@@ -40,7 +40,7 @@ impl Results {
 }
 
 impl SIR {
-    pub fn new(S: f32, I: f32, R: f32) -> Self {
+    pub fn new(S: f64, I: f64, R: f64) -> Self {
         return SIR {
             S,
             I,
@@ -49,21 +49,19 @@ impl SIR {
     }
 }
 
-fn Sdot(state: &SIR, params: &Parameters , i: &usize) -> f32 {
-    - params.Beta[*i] * (( (state.S * state.I) / (state.S + state.I + state.R) ) as f32) +
-    params.Alpha[*i] * (state.R as f32)
+fn Sdot(state: &SIR, params: &Parameters , i: &usize) -> f64 {
+    (- params.Beta[*i] * ( (state.S * state.I) / (state.S + state.I + state.R) ) + params.Alpha[*i] * state.R)
 }
 
-fn Idot(state: &SIR, params: &Parameters , i: &usize) -> f32 {
-    params.Beta[*i] * (( (state.S * state.I) / (state.S + state.I + state.R) ) as f32) -
-    params.Gamma[*i] * ( state.I as f32)
+fn Idot(state: &SIR, params: &Parameters , i: &usize) -> f64 {
+    (params.Beta[*i] * ( (state.S * state.I) / (state.S + state.I + state.R) ) - params.Gamma[*i] * ( state.I))
 }
 
-fn Rdot(state: &SIR, params: &Parameters , i: &usize) -> f32 {
-    params.Gamma[*i] * ( state.I as f32) - params.Alpha[*i] * (state.R as f32)
+fn Rdot(state: &SIR, params: &Parameters , i: &usize) -> f64 {
+    (params.Gamma[*i] * state.I - params.Alpha[*i] * state.R)
 }
 
-fn step(state: &SIR, params: &Parameters, i: &usize, dt: &f32) -> SIR {
+fn step(state: &SIR, params: &Parameters, i: &usize, dt: &f64) -> SIR {
     let S = state.S + Sdot(state, params, i) * dt;
     let I = state.I + Idot(state, params, i) * dt;
     let R = state.R + Rdot(state, params, i) * dt;
@@ -78,9 +76,9 @@ fn capture_state(state: &SIR, results: &mut Results) {
 }
 
 
-pub fn run(runtime: f32, dt: f32, initial_state: SIR, params: Parameters) -> Results {
+pub fn run(runtime: f64, dt: f64, initial_state: SIR, params: Parameters) -> Results {
     let mut i = 0 as usize;
-    let mut t = 0 as f32;
+    let mut t = 0 as f64;
     let mut state = initial_state;
     let mut res = Results::new();
     while t < runtime - dt {
